+++ src/main/java/com/lowdragmc/lowdraglib/utils/FacadeBlockWorld.java
@@ -26,14 +26,12 @@ import javax.annotation.ParametersAreNonnullByDefault;
 @OnlyIn(Dist.CLIENT)
 public class FacadeBlockWorld extends DummyWorld {
 
-    public final Level world;
     public final BlockPos pos;
     public final BlockState state;
     public final BlockEntity tile;
 
     public FacadeBlockWorld(Level world, BlockPos pos, BlockState state, BlockEntity tile) {
         super(world);
-        this.world = world;
         this.pos = pos;
         this.state = state;
         this.tile = tile;
@@ -42,65 +40,60 @@ public class FacadeBlockWorld extends DummyWorld {
     @Nullable
     @Override
     public BlockEntity getBlockEntity(@Nonnull BlockPos pPos) {
-        return world == null ? super.getBlockEntity(pPos) : pPos.equals(pos) ? tile : world.getBlockEntity(pPos);
+        return pPos.equals(pos) ? tile : getLevel().getBlockEntity(pPos);
     }
 
     @Override
     @ParametersAreNonnullByDefault
     @Nonnull
     public BlockState getBlockState(BlockPos pPos) {
-        return  world == null ? super.getBlockState(pPos) : pPos.equals(pos) ? state : world.getBlockState(pPos);
+        return pPos.equals(pos) ? state : getLevel().getBlockState(pPos);
     }
 
     @Override
     @Nonnull
     public LevelLightEngine getLightEngine() {
-        return world == null ? super.getLightEngine() : world.getLightEngine();
+        return getLevel().getLightEngine();
     }
 
     @Override
     public int getBrightness(@Nonnull LightLayer lightType, @Nonnull BlockPos pos) {
-        return  world == null ? super.getBrightness(lightType, pos) : world.getBrightness(lightType, pos);
+        return getLevel().getBrightness(lightType, pos);
     }
 
     @Override
     public int getBlockTint(@Nonnull BlockPos blockPos, @Nonnull ColorResolver colorResolver) {
-        return  world == null ? super.getBlockTint(blockPos, colorResolver) : world.getBlockTint(blockPos, colorResolver);
+        return getLevel().getBlockTint(blockPos, colorResolver);
     }
 
     @Override
     public boolean canSeeSky(@Nonnull BlockPos pos) {
-        return  world == null ? super.canSeeSky(pos) : world.canSeeSky(pos);
+        return getLevel().canSeeSky(pos);
     }
 
     @Nonnull
     @Override
     public DimensionType dimensionType() {
-        return  world == null ? super.dimensionType() : world.dimensionType();
+        return getLevel().dimensionType();
     }
 
     @Override
     public boolean isEmptyBlock(@Nonnull BlockPos pPos) {
-        return  world == null ? super.isEmptyBlock(pPos) : !pPos.equals(pos) && world.isEmptyBlock(pPos);
+        return !pPos.equals(pos) && getLevel().isEmptyBlock(pPos);
     }
 
     @Nonnull
     @Override
     @OnlyIn(Dist.CLIENT)
     public Holder<Biome> getBiome(@Nonnull BlockPos pos) {
-        return  world == null ? super.getBiome(pos) : world.getBiome(pos);
+        return getLevel().getBiome(pos);
     }
-
-//    @Override
-//    public @Nullable BlockEntity getExistingBlockEntity(BlockPos pos) {
-//        return  world == null ? super.getExistingBlockEntity(pos) : pos.equals(this.pos) ? tile : world.getExistingBlockEntity(pos);
-//    }
 
     @Override
     @OnlyIn(Dist.CLIENT)
     public void setParticleManager(@Nonnull ParticleManager particleManager) {
         super.setParticleManager(particleManager);
-        if (world instanceof DummyWorld dummyWorld) {
+        if (getLevel() instanceof DummyWorld dummyWorld) {
             dummyWorld.setParticleManager(particleManager);
         }
     }
@@ -110,7 +103,7 @@ public class FacadeBlockWorld extends DummyWorld {
     @OnlyIn(Dist.CLIENT)
     public ParticleManager getParticleManager() {
         ParticleManager particleManager = super.getParticleManager();
-        if (particleManager == null && world instanceof DummyWorld dummyWorld) {
+        if (particleManager == null && getLevel() instanceof DummyWorld dummyWorld) {
             return dummyWorld.getParticleManager();
         }
         return particleManager;